requirements.txt
undetected-chromedriver==3.5.4
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
python-dotenv==1.0.0